+++ CHANGELOG.md
@@ -9,6 +9,8 @@ resources:
   * Removed the WindowsServiceAccount parameter that was depricated in v3.1
 * SPUserProfileServiceApp
   * Changed the MySiteHostLocation parameter to a required parameter
+* SPWebAppAuthentication
+  * Updated resource to add support for Basic Authentication
 
 ## UNRELEASED